chore: Refactor uint64 module with external bellpepper-uint64 package
- Replaced the local `uint64` modules with the `bellpepper_uint64` package in the `keccak` and `sha512` crates.
- Removed `UInt64` struct and associated functions/tests from the `keccak` & `sha512` crates.
- Created a `uint64` crate with details including license (both MIT and Apache 2.0), authors, description, and dependencies.
- Updated `Cargo.toml` files for the `sha512` and `keccak` crates, adding `bellpepper-uint64` as a new dependency, and for the project, including `uint64` in the project members.
- Created comprehensive README and license files for the new `uint64` crate.
