+++ crates/sha512/Cargo.toml
@@ -13,6 +13,7 @@ repository.workspace = true
 bellpepper-core = { workspace = true }
 bellpepper = { workspace = true }
 ff = { workspace = true }
+bellpepper-uint64 = { version = "0.1.0", path = "../uint64"}
 
 [dev-dependencies]
 blstrs = "0.7.0"
