+++ crates/sha512/src/lib.rs
@@ -1,2 +1,1 @@
 pub mod sha512;
-pub mod uint64;
